feat: Add DirectWriteAbsTime time sync procedure for non-conformant outstations
Add support for time synchronization via direct write of Group 50 Var 1
without delay measurement. This is a workaround for outstations that do
not support the DELAY_MEASURE function code.

This method does not account for network latency and should only be used
when standard time synchronization procedures are not supported.

Closes #374